Tambahan var 1 dan 2

+++ src/rabu.kt
@@ -1,4 +1,4 @@
-fun factorial(n: Int): Long {
+fun factorial(n: Int): Int {
     if (n == 0) {
         return 1
     } else {
